+++ include/UdpTransport.hpp
@@ -8,7 +8,8 @@
 
 class UdpTransport : public ITransport {
 public:
-    UdpTransport(std::string host, int32_t port) : socket_{std::move(host), port} {}
+
+    UdpTransport(std::string host, uint16_t port) : socket_{std::move(host), port} {}
 
     void connect() override                         { socket_.connect(); }
     std::size_t sendString(const std::string& s) override { return socket_.sendString(s); }
